global styles and user interaction
